+++ jac/jaclang/compiler/passes/main/inheritance_pass.py
@@ -6,7 +6,7 @@ from typing import Optional
 
 import jaclang.compiler.unitree as uni
 from jaclang.compiler.passes import AstPass
-from jaclang.compiler.symtable import Symbol, UniScopeNode
+from jaclang.compiler.unitree import Symbol, UniScopeNode
 from jaclang.settings import settings
